+++ src/Frey.js
@@ -63,8 +63,14 @@ class Frey extends Base {
     }
 
     if (filteredChain.indexOf('prepare') < 0 && (startAt < 0 || startAt > _.findIndex(chain, {name: 'prepare'}))) {
-      if (cmd !== 'remote') {
+      if (cmd !== 'remote' && cmd !== 'convert') {
         filteredChain.unshift('prepare')
+      }
+    }
+
+    if (filteredChain.indexOf('format') < 0 && (startAt < 0 || startAt > _.findIndex(chain, {name: 'format'}))) {
+      if (cmd === 'convert' && !cliargs.bail) {
+        filteredChain.push('format')
       }
     }